Handle slashed in bb-build's --output
As far as I understand, the current semantics of the --output flag is to
expect an executable name (without any slash) and the generated
executable would be copied to scripts/bin/

This seems counter intuitive to me. Instead I expected the following
command:

> bb-build --input something --output /path/to/blah.exe

to generate an executable named blah.exe in the directory /path/to.
This patch implement this semantics.